+++ environments/gloo-gateway/gateway-api/1.17/additional-cluster-1/gloo-agent/test.sh
@@ -4,7 +4,7 @@
 if [[ ${gloo_mesh_version} == "" ]]
   then
     # provide gloo_mesh_version variable
-    echo "Please provide the gloo_mesh_version to use (i.e. v2.8.0-beta0-2025-01-29-main-4bf61e4969):"
+    echo "Please provide the gloo_mesh_version to use (i.e. v2.8.0-beta0-2025-01-30-jmcguire-configure-grpc-max-size-on-analyzer-96ebf69cbd):"
     read gloo_mesh_version
 fi
 
@@ -105,7 +105,7 @@ spec:
             enabled: true
 
     repoURL: https://storage.googleapis.com/gloo-platform-dev/platform-charts/helm-charts
-    targetRevision: v2.8.0-beta0-2025-01-29-main-4bf61e4969
+    targetRevision: v2.8.0-beta0-2025-01-30-jmcguire-configure-grpc-max-size-on-analyzer-96ebf69cbd
   syncPolicy:
     automated:
       prune: true
@@ -163,7 +163,7 @@ spec:
             skipVerify: true
 
     repoURL: https://storage.googleapis.com/gloo-platform-dev/platform-charts/helm-charts
-    targetRevision: v2.8.0-beta0-2025-01-29-main-4bf61e4969
+    targetRevision: v2.8.0-beta0-2025-01-30-jmcguire-configure-grpc-max-size-on-analyzer-96ebf69cbd
   syncPolicy:
     automated:
       prune: true